feat(image-upload): Fixing image upload to save file AWS S3 - fixed image upload to accept files instead of base64 string - ensured that all test pass [Delivers #163725863]
